+++ src/test/java/calculator/TestModulo.java
@@ -1,0 +1,85 @@
+package calculator;
+
+//Import Junit5 libraries for unit testing:
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.*;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public class TestModulo {
+
+    private final int value1 = 8;
+    private final int value2 = 6;
+    private final int zero = 0;
+    private final int negativeValue = -3;
+    private Modulo op;
+    private List<Expression> params;
+
+    @BeforeEach
+    void setUp() {
+        params = Arrays.asList(new MyNumber(value1), new MyNumber(value2));
+        try {
+            op = new Modulo(params);
+        } catch (IllegalConstruction e) {
+            fail();
+        }
+    }
+
+    @Test
+    void testConstructor1() {
+        // It should not be possible to create an expression without null parameter list
+        assertThrows(IllegalConstruction.class, () -> op = new Modulo(null));
+    }
+
+    @SuppressWarnings("AssertBetweenInconvertibleTypes")
+    @Test
+    void testConstructor2() {
+        // A Minus expression should not be the same as a Modulo expression
+        try {
+            assertNotSame(op, new Minus(new ArrayList<>()));
+        } catch (IllegalConstruction e) {
+            fail();
+        }
+    }
+
+    @Test
+    void testEquals() {
+        // Two similar expressions, constructed separately (and using different constructors) should not be equal
+        List<Expression> p = Arrays.asList(new MyNumber(value1), new MyNumber(value2));
+        try {
+            Modulo e = new Modulo(p, Notation.INFIX);
+            assertEquals(op, e);
+        } catch (IllegalConstruction e) {
+            fail();
+        }
+    }
+
+    @SuppressWarnings("ConstantConditions")
+    @Test
+    void testNull() {
+        assertDoesNotThrow(() -> op == null); // Direct way to to test if the null case is handled.
+    }
+
+    @Test
+    void testHashCode() {
+        // Two similar expressions, constructed separately (and using different constructors) should have the same hashcode
+        List<Expression> p = Arrays.asList(new MyNumber(value1), new MyNumber(value2));
+        try {
+            Modulo e = new Modulo(p, Notation.INFIX);
+            assertEquals(e.hashCode(), op.hashCode());
+        } catch (IllegalConstruction e) {
+            fail();
+        }
+    }
+
+    @Test
+    void testNullParamList() {
+        params = null;
+        assertThrows(IllegalConstruction.class, () -> op = new Modulo(params));
+    }
+
+
+
+}